search: refactor search class into several classes
makes the code a bit more managable
